added middle mouse
