+++ src/com/jagex/runescape/util/ChatEncoder.java
@@ -6,10 +6,10 @@ public class ChatEncoder {
 
 	public static char[] message = new char[100];
 	public static Buffer messageBuffer = new Buffer(new byte[100]);
-	public static char VALID_CHARACTERS[] = { ' ', 'e', 't', 'a', 'o', 'i', 'h', 'n', 's', 'r', 'd', 'l', 'u', 'm', 'w',
+	public static char VALID_CHARACTERS[] = {' ', 'e', 't', 'a', 'o', 'i', 'h', 'n', 's', 'r', 'd', 'l', 'u', 'm', 'w',
 			'c', 'y', 'f', 'g', 'p', 'b', 'v', 'k', 'x', 'j', 'q', 'z', '0', '1', '2', '3', '4', '5', '6', '7', '8',
 			'9', ' ', '!', '?', '.', ',', ':', ';', '(', ')', '-', '&', '*', '\\', '\'', '@', '#', '+', '=', '\243',
-			'$', '%', '"', '[', ']' };
+			'$', '%', '"', '[', ']', 'æ', 'ø', 'å' };
 
 	public static String get(int length, Buffer buffer) {
 		int count = 0;
